+++ python/ray/tests/test_object_manager.py
@@ -11,6 +11,7 @@ import time
 import warnings
 
 import ray
+from ray import ray_constants
 from ray.cluster_utils import Cluster
 
 # TODO(yuhguo): This test file requires a lot of CPU/memory, and
@@ -45,6 +46,7 @@ def ray_start_cluster_with_resource():
 
 # This test is here to make sure that when we broadcast an object to a bunch of
 # machines, we don't have too many excess object transfers.
+@pytest.mark.skipif(ray_constants.direct_call_enabled(), reason="TODO(ekl)")
 def test_object_broadcast(ray_start_cluster_with_resource):
     cluster, num_nodes = ray_start_cluster_with_resource
 
@@ -235,7 +237,7 @@ def test_object_transfer_retry(ray_start_cluster):
     # Transfer an object to warm up the object manager.
     ray.get(f.remote(10**6))
 
-    x_ids = [f.remote(10**i) for i in [1, 2, 3, 4]]
+    x_ids = [f.remote(10**i) for i in [6]]
     assert not any(
         ray.worker.global_worker.core_worker.object_exists(x_id)
         for x_id in x_ids)
