+++ python/ray/tests/test_signal.py
@@ -1,5 +1,7 @@
+import pytest
 import time
 
+from ray import ray_constants
 import ray
 import ray.experimental.signal as signal
 
@@ -274,6 +276,9 @@ def test_forget(ray_start_regular):
     assert len(result_list) == count
 
 
+@pytest.mark.skipif(
+    ray_constants.direct_call_enabled(),
+    reason="TODO(ekl): this requires reconstruction")
 def test_signal_on_node_failure(two_node_cluster):
     """Test actor checkpointing on a remote node."""
