+++ python/ray/tune/tests/test_cluster.py
@@ -123,6 +123,8 @@ def test_trial_processed_after_node_failure(start_connected_emptyhead_cluster):
 
     cluster.remove_node(node)
     runner.step()
+    if not mock_process_failure.called:
+        runner.step()
     assert mock_process_failure.called
 
 
@@ -259,11 +261,9 @@ def test_trial_migration(start_connected_emptyhead_cluster):
     cluster.remove_node(node2)
     cluster.wait_for_nodes()
     runner.step()  # Recovery step
-    assert t2.last_result["training_iteration"] == 2
-    for i in range(1):
+    if t2.status != Trial.TERMINATED:
         runner.step()
-
-    assert t2.status == Trial.TERMINATED
+    assert t2.status == Trial.TERMINATED, runner.debug_string()
 
     # Test recovery of trial that won't be checkpointed
     t3 = Trial("__fake", **{"stopping_criterion": {"training_iteration": 3}})
@@ -274,7 +274,9 @@ def test_trial_migration(start_connected_emptyhead_cluster):
     cluster.remove_node(node3)
     cluster.wait_for_nodes()
     runner.step()  # Error handling step
-    assert t3.status == Trial.ERROR
+    if t3.status != Trial.ERROR:
+        runner.step()
+    assert t3.status == Trial.ERROR, runner.debug_string()
 
     with pytest.raises(TuneError):
         runner.step()
@@ -340,9 +342,9 @@ def test_migration_checkpoint_removal(start_connected_emptyhead_cluster):
 
     runner.step()  # Recovery step
     for i in range(3):
-        runner.step()
-
-    assert t1.status == Trial.TERMINATED
+        if t1.status != Trial.TERMINATED:
+            runner.step()
+    assert t1.status == Trial.TERMINATED, runner.debug_string()
 
 
 def test_cluster_down_simple(start_connected_cluster, tmpdir):
